Clarify snapshot test fixture ownership
